Fixed the unit tests, cleaned up some other code and added formatted print statements
I fixed the issues I was having with the BorrowUC_CTL test class due to
the panels being called in the methods. I also added some formatted
println statements to the MemberMapDAO test class to further confirm
that the methods are operating as expected. This may not be required but
I prefer them.